offscreen-canvas/pixel-manipulation/2d.imageData.object.undefined.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.imageData.object.undefined</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>

<h1>2d.imageData.object.undefined</h1>
<p class="desc">ImageData.data converts undefined to 0</p>


<script>
promise_test(function(t) {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var imgdata = ctx.getImageData(0, 0, 10, 10);
imgdata.data[0] = 100;
imgdata.data[0] = undefined;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");

return Promise.resolve();
}, "ImageData.data converts undefined to 0");
</script>
